prevent nexttrace from being started repeatedly

+++ MainForm.cs
@@ -30,6 +30,7 @@ namespace OpenTrace
         private Button startTracerouteButton;
         private bool gridResizing = false;
         private bool appForceExiting = false;
+        private bool enterPressed = false;
 
         private ExceptionalOutputForm exceptionalOutputForm = new ExceptionalOutputForm();
 
@@ -73,6 +74,7 @@ namespace OpenTrace
             // 创建控件
             HostInputBox = new ComboBox { Text = "" };
             HostInputBox.KeyDown += HostInputBox_KeyDown;
+            HostInputBox.KeyUp += HostInputBox_KeyUp;
             HostInputBox.TextChanged += HostInputBox_TextChanged;
             if(UserSettings.traceHistory != null || UserSettings.traceHistory!= "")
             {
@@ -210,9 +212,16 @@ namespace OpenTrace
 
         private void HostInputBox_KeyDown(object sender, KeyEventArgs e)
         {
-            if (e.Key == Keys.Enter)
+            if (e.Key == Keys.Enter && !enterPressed)
             {
-                if (CurrentInstance != null)  StopTraceroute();
+                enterPressed = true;
+            }
+        }
+        private void HostInputBox_KeyUp(object sender, KeyEventArgs e)
+        {
+            if (e.Key == Keys.Enter && enterPressed)
+            {
+                enterPressed = false;
                 StartTracerouteButton_Click(sender, e);
             }
         }
@@ -224,6 +233,7 @@ namespace OpenTrace
 
         private void StartTracerouteButton_Click(object sender, EventArgs e)
         {
+            
             if(protocolSelection.SelectedValue.ToString() != "ICMP" && RuntimeInformation.IsOSPlatform(OSPlatform.Windows))
             {
                 MessageBox.Show(Resources.WINDOWS_TCP_UDP_UNSUPPORTED);
@@ -234,13 +244,9 @@ namespace OpenTrace
                 StopTraceroute();
                 return;
             }
-            tracerouteResultCollection.Clear(); // 清空原有GridView
-            ResetMap(); // 重置地图
-            Title = Resources.APPTITLE;
-            NextTraceWrapper instance;
             try
             {
-                instance = new NextTraceWrapper();
+                CurrentInstance = new NextTraceWrapper();
             }
             catch (FileNotFoundException)
             {
@@ -260,6 +266,9 @@ namespace OpenTrace
                 return;
             }
 
+            tracerouteResultCollection.Clear(); // 清空原有GridView
+            ResetMap(); // 重置地图
+            Title = Resources.APPTITLE;
             // 处理文本框输入
             string readyToUseIP;
             if (ResolvedIPSelection.Visible == true && ResolvedIPSelection.SelectedIndex != 0)
@@ -344,36 +353,29 @@ namespace OpenTrace
             }
             UserSettings.traceHistory = String.Join("\n", HostInputBox.Items.Select(item => item.Text));
             UserSettings.SaveSettings();
-            CurrentInstance = instance;
+
             startTracerouteButton.Text = Resources.STOP;
 
             // 处理NextTrace实例发回的结果
-            instance.Output.CollectionChanged += Instance_OutputCollectionChanged;
-            instance.ExceptionalOutput += Instance_ExceptionalOutput;
-            instance.AppQuit += Instance_AppQuit;
-            instance.Run(readyToUseIP, (bool)MTRMode.Checked, dataProviderSelection.SelectedKey, protocolSelection.SelectedKey);
+            CurrentInstance.Output.CollectionChanged += Instance_OutputCollectionChanged;
+            CurrentInstance.ExceptionalOutput += Instance_ExceptionalOutput;
+            CurrentInstance.AppQuit += Instance_AppQuit;
+            CurrentInstance.Run(readyToUseIP, (bool)MTRMode.Checked, dataProviderSelection.SelectedKey, protocolSelection.SelectedKey);
             
         }
-
         private void Instance_AppQuit(object sender, AppQuitEventArgs e)
         {
             Application.Instance.InvokeAsync(() =>
             {
-                if (appForceExiting != true)
+                CurrentInstance = null;
+                startTracerouteButton.Text = Resources.START;
+                if (appForceExiting != true && (e.ExitCode != 0))
                 {
-                    // 主动结束
-                    startTracerouteButton.Text = Resources.START;
-                    CurrentInstance = null;
-                    if (e.ExitCode != 0)
-                    {
-                        MessageBox.Show(Resources.EXCEPTIONAL_EXIT_MSG + e.ExitCode, MessageBoxType.Warning);
-                    }
+                    // 主动结束，退出代码不为 0 则证明有异常
+                    MessageBox.Show(Resources.EXCEPTIONAL_EXIT_MSG + e.ExitCode, MessageBoxType.Warning);
                 }
-                else
-                {
-                    // 强制结束
-                    appForceExiting = false;
-                }
+                // 强制结束一般退出代码不为 0，不提示异常。
+                appForceExiting = false;
             });
         }
 
@@ -396,35 +398,32 @@ namespace OpenTrace
             {
                 Application.Instance.InvokeAsync(() =>
                 {
-                    int HopNo = int.Parse(((TracerouteResult)e.NewItems[0]).No);
-                    if (HopNo > tracerouteResultCollection.Count)
+                    try
                     {
-                        // 正常添加新的跳
-                        tracerouteResultCollection.Add(new TracerouteHop((TracerouteResult)e.NewItems[0]));
-                        UpdateMap((TracerouteResult)e.NewItems[0]);
-                        tracerouteGridView.ScrollToRow(tracerouteResultCollection.Count - 1);
-                    }
-                    else
-                    {
-                        // 修改现有的跳
-                        tracerouteResultCollection[HopNo - 1].HopData.Add((TracerouteResult)e.NewItems[0]);
-                        tracerouteGridView.ReloadData(HopNo - 1);
-                    }
+                        int HopNo = int.Parse(((TracerouteResult)e.NewItems[0]).No);
+                        if (HopNo > tracerouteResultCollection.Count)
+                        {
+                            // 正常添加新的跳
+                            tracerouteResultCollection.Add(new TracerouteHop((TracerouteResult)e.NewItems[0]));
+                            UpdateMap((TracerouteResult)e.NewItems[0]);
+                            tracerouteGridView.ScrollToRow(tracerouteResultCollection.Count - 1);
+                        }
+                        else
+                        {
+                            // 修改现有的跳
+                            tracerouteResultCollection[HopNo - 1].HopData.Add((TracerouteResult)e.NewItems[0]);
+                            tracerouteGridView.ReloadData(HopNo - 1);
+                        }
+                    } catch { }
                 });
             }
         }
 
         private void StopTraceroute()
         {
-            if(CurrentInstance != null)
-            {
+            if (CurrentInstance != null && !appForceExiting) {
                 appForceExiting = true;
                 CurrentInstance.Kill();
-                startTracerouteButton.Text = Resources.START;
-                CurrentInstance.Output.CollectionChanged -= Instance_OutputCollectionChanged;
-                CurrentInstance.AppQuit -= Instance_AppQuit;
-                CurrentInstance.ExceptionalOutput -= Instance_ExceptionalOutput;
-                CurrentInstance = null;
             }
         }
